requirements.txt
# Web 框架
Flask==3.0.2

# Nacos SDK
nacos-sdk-python==0.1.12

# HTTP 请求库
requests==2.31.0

# 其他依赖
Werkzeug==3.0.1
click==8.1.7
urllib3==2.2.1
certifi==2024.2.2
charset-normalizer==3.3.2
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.5 